Adicionando pagina sign-out e animação lottie

+++ projeto-recados/js/signup.js
@@ -1,6 +1,5 @@
 let formCadastro = document.querySelector("#formulario-cadastro");
 
-
 let label_email = document.querySelector("#label-email");
 let input_email = document.querySelector("#input-email");
 
@@ -10,21 +9,16 @@ let input_senha = document.querySelector("#input-senha");
 let label_conf_senha = document.querySelector("#label-confirma-senha");
 let input_conf_senha = document.querySelector("#input-confirma-senha");
 
-let formAcesso = document.querySelector('.formulario');
+let formAcesso = document.querySelector(".formulario");
 let acesso_email = document.querySelector("#input-acesso-email");
 
 let acesso_senha = document.querySelector("#input-acesso-senha");
-
-
-
-
 
 //listas e objetos
 
 //Eventos
 formAcesso.addEventListener("submit", (e) => {
   e.preventDefault();
-
 });
 
 formCadastro.addEventListener("submit", (e) => {
@@ -32,20 +26,14 @@ formCadastro.addEventListener("submit", (e) => {
   enviardados();
 });
 
-
-
-
 input_email.addEventListener("keyup", confirmaEmail);
 input_senha.addEventListener("keyup", confirmasenha);
 input_conf_senha.addEventListener("keyup", validasenha);
 
-
-
-
-
 //REGRAS REGEX PARA VALIDAÇÃO SENHA
 //Mínimo de oito caracteres, pelo menos uma letra maiúscula, uma letra minúscula, um número e um caractere especial:
-let regSenha = /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$/;
+let regSenha =
+  /^(?=.*[a-z])(?=.*[A-Z])(?=.*\d)(?=.*[@$!%*?&])[A-Za-z\d@$!%*?&]{8,}$/;
 
 // funções
 
@@ -67,13 +55,12 @@ function confirmasenha() {
     label_senha.setAttribute("style", "color: red");
     label_senha.innerHTML = "Senha: mínimo de 8 caracteres";
     input_senha.setAttribute("style", "border-color: red");
-  } 
-  else if(senhaValida === null){ 
-    label_senha.setAttribute("style", "color: red");       
-    label_senha.innerHTML = 'Senha: *Deve conter uma letra maiuscula e caracter';
-    input_senha.setAttribute("style", "border-color: red");  
-  }
-  else {
+  } else if (senhaValida === null) {
+    label_senha.setAttribute("style", "color: red");
+    label_senha.innerHTML =
+      "Senha: Deve conter uma letra maiuscula e caracter especial";
+    input_senha.setAttribute("style", "border-color: red");
+  } else {
     label_senha.setAttribute("style", "color: green");
     label_senha.innerHTML = "Senha:";
     input_senha.setAttribute("style", "border-color: green");
@@ -93,7 +80,6 @@ function validasenha() {
   }
 }
 
-
 function enviardados() {
   if (
     input_email.value === "" ||
@@ -101,45 +87,55 @@ function enviardados() {
     input_senha.value === "" ||
     input_senha.value.length < 8 ||
     input_conf_senha.value === "" ||
-    input_senha.value !== input_conf_senha.value) {
+    input_senha.value !== input_conf_senha.value
+  ) {
     alert("Verifique se os campos estão preenchidos da forma correta!");
     return;
   } else {
+    
     salvardados();
     limparDados();
-    var dialog = confirm("Dados criados com sucesso, retornar para pagina de acesso?");
-        if (dialog) {
-          window.location.href = "index.html";
-        }
-        else {
-          window.location.href = "signup.html";
-        }
-    }    
   }
-
-
-function salvardados(){
-  
-  let dados={
-    usuario : input_email.value,
-    senha : input_senha.value,
-    recados:[]
-  }
-
-  let pegadados = JSON.parse(localStorage.getItem('dados-usuario') || '[]');
-  pegadados.unshift(dados)
-  const userString = JSON.stringify(pegadados);
-  localStorage.setItem("dados-usuario",userString);
-  // sessionStorage.setItem('login',userString);
 }
 
-function limparDados(){
-  document.querySelector('#input-email').value ='';
-  document.querySelector('#input-senha').value ='';
-  document.querySelector('#input-confirma-senha').value ='';
-  
+function salvardados() {
+  let existeDados = false;
+  let dados = {
+    usuario: input_email.value,
+    senha: input_senha.value,
+    recados: [],
+  };
+
+  let pegadados = JSON.parse(localStorage.getItem("dados-usuario") || "[]");
+
+  for (const indice in pegadados) {
+    if (pegadados[indice].usuario === dados.usuario) {
+      var  aviso = alert("Usuário já existe!");
+      existeDados = true;
+      return aviso;
+    }
+  }
+  if (existeDados !== true) {
+    pegadados.unshift(dados);
+    const userString = JSON.stringify(pegadados);
+    localStorage.setItem("dados-usuario", userString);
+    var  aviso =alert(
+      "Dados criados com sucesso!"
+    );
+    window.location.href = "index.html";
+    return aviso;
+  }
+  return existeDados;
+}
+
+function limparDados() {
+  document.querySelector("#input-email").value = "";
+  document.querySelector("#input-senha").value = "";
+  document.querySelector("#input-confirma-senha").value = "";
+
   label_email.setAttribute("style", "color: black");
   input_email.setAttribute("style", "border-color: black");
+  input_email.focus();
 
   label_senha.setAttribute("style", "color: black");
   input_senha.setAttribute("style", "border-color: black");
@@ -147,7 +143,3 @@ function limparDados(){
   label_conf_senha.setAttribute("style", "color: black");
   input_conf_senha.setAttribute("style", "border-color: black");
 }
-
-
-
-
